sway/config: Use wob for sound and brightness overlays
wob is a Wayland overlay bar for things like sound and brightness
levels. It expects values to be between 0 and 100, which means we
need to use light instead of brightnessctl because the brightness
level is in lumens or something so it goes up to 1060 on my Think
Pad Carbon X1 Gen 6 and light represents it as a percentage. Also
we need to use pamixer instead of pactl because there is no simple
way to get the current volume with pactl.

See: https://github.com/francma/wob